requirements.txt
click==8.1.3
Flask==2.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.6
scikit-learn==1.0.2
scipy==1.9.3
six==1.16.0
SQLAlchemy==1.4.40
threadpoolctl==3.1.0
Werkzeug==2.0.1
WTForms==3.0.1
gunicorn
urllib3